Add fuzzy file open support
While opening buffers from a list may be the more common operation,
sometimes it is necessary to open a file that has not been opened
before.
This change adds support for doing so. We try to find the root directory
of the source tree we are working on, represented as the directory
common to all already opened files. We then use 'fd' to list all files
below that directory and filter them in a fuzzy manner using fzy.