Do not search files based on single "root" directory
The approach we took for searching files so far was to deduce the "root"
directory of all files edited and search the directory tree below that.
While that is conceptually feasible and correct, it can result in a blow
up of the search space. If, for example, the user is working on a
project below his/her home directory, but he/she happens to have opened
some system file, chances are the root directory is inferred to be /.
Now, while 'fd', the program we use for listing files is fast, at least
on a spinning disk it is far from a great user experience to traverse
the entire file system.
This change reworks the logic to reduce the search space. Luckily, 'fd'
accept multiple directories as input, and so with this change we simply
supply all the directories containing files that are opened as buffers.
That typically results in a reasonable search space while in most cases
covering all the files a user would want to open.