Make the file finder and the fuzzy searcher configurable
Up until now we have used fixed values for the fuzzy search program to
use ('fzy') as well as the command to list files below a set of
directories ('fd'). That is not a very user friendly way, as different
users will have different preferences and/or requirements.
With this change we make the file finder as well as the fuzzy searcher
configurable through global variables. Note that the values are only
read once from the Neovim instance and then cached subsequently.