Use Neovim API for listing buffers
For an unknown reason the ':ls' command sometimes (every time except for
the first invocation?) jumps into (interactive) pager mode when a lot of
buffers are open (exceeding one screen page). That is an obvious problem
as it breaks the plugin's work flow which is designed to be
non-interactive.
With this change we switch to using the nvim_list_bufs API which does
not have this problem and requires no parsing of output. In addition to
solving this issue, it also does not temporarily display the output in
the window, which seemingly was the case beforehand.