Fix error reporting for pipeline execution failures
As it turns out there were a couple of issues with the process pipeline
we use in order to implement the fuzzy file search functionality. These
have all been fixed.
First, errors were not reported at all. Popen's communicate does not
raise any CalledProcessError exceptions and so errors were never
flagged.
Second, we did not detect failures of the first of the two commands
(well, that's sort of a sub-issue of not detecting any errors, but it is
important to be clear about). In essence, we have behavior similar to
Bash without the pipefail option set. That is a problem because errors
hit during the search for files were never reported.
Lastly, even if such failures would have been detected for the first
process, we would still need to make sure to terminate the second
process. Otherwise this process would stay activate for a potentially
long time.