Filter out non-existent and subsumed directories
When using fd as the file listing tool we have a problem when a buffer
references a directory that currently or no longer exists, as the
program will just error out complaining about this fact.
With this change we solve this problem by only searching existent
directories. We also go one step further and remove all directories
subsuming others from the list passed to the program. That can be useful
for sessions where a lot of buffers are open in a lot of different
directories, because we could possibly exceed the maximum length of
command line arguments.