Fix relative path handling of removeSubsumedPaths
When the removeSubsumedPaths function is provided with a list of
relative paths the result may not be as expected, if they contains
leading dots (e.g., ./<relative-path>). The reason is that the
os.path.commonpath function removes such prefixes.
This change fixes these problems by normalizing the result to the
absolute path. Note that doing so is insufficient for fixing the general
case of input containing a mixture of ./<relative-path> and
<relative-path> paths. Currently we assume that clients do one thing or
another.