Use vim.api.list_bufs instead of vim.request
As it turns out the vim.api module contains "native" bindings for a
bunch of functionality, including the nvim_list_bufs, which we use. With
them, instead of making a "generic" request to the API, we get a neater
looking function call syntax with a shortened function name.
This change switches our call over to using vim.api.list_bufs.