Use Neovim's working directory for finder
We have seen problems where nfuzz effectively got into an unrecoverable
state when used with fd-find the directory in which the program was
started was removed. The reason for that behavior was that fd-find would
always fail, complaining that the current directory was no longer
present.
This patch fixes this issue via two changes. First, we work with
Neovim's notion of a working directory. Depending on how the editor is
configured, that working directory may change when activating a
different buffer. In any case, it can be changed at runtime. Second, we
guarantee that the finder is invoked with the current working directory
as a first argument, while recommending that the user uses a finder that
can honor this directory instead of inferring the owning processes
working directory. For fd-find, for example, that can be accomplished
via the --base-directory option.