Show buffer number during buffer selection
So far we have always selected buffers based on their names only. In
most cases that worked great and was what the user expected. However,
sometimes a buffer's number may be known and "less ambiguous" to the
user, e.g., in directory hierarchies with multiple files of the same
name.
With this change we include the buffer number in the user-facing buffer
selection. This way, the user is free to include said number in the
fuzzy search. We opted for showing the number after the buffer name
itself. The idea here is that this helps disambiguate a fuzzy buffer
name search, by just appending one or more digits of the number.