Handle unnamed buffers gracefully
Buffer listings may contain buffers that do not have a name. Quick fix
windows are such candidates. In the past, it was impossible for us to
handle these unnamed buffers correctly, because it was the name that we
used for opening them. As a result, not only were these buffers
unselectable, they also caused annoying runtime errors, because Python's
os.path.relpath function does not handle empty input paths.
Given that we now switched to working with buffer numbers instead, we
are able to handle such unnamed buffers more gracefully. With this
change we make them show up as '<unnamed>' in the buffer listing, while
still having their buffer number present, allowing for selection based
on said number.